Start the big refactor
